Move utils.make_pretty_large_file_content into pretty_printers module
