+++ onlinejudge_command/output_comparators.py
@@ -8,6 +8,7 @@ import tempfile
 from logging import getLogger
 from typing import *
 
+import onlinejudge_command.pretty_printers as pretty_printers
 import onlinejudge_command.utils as utils
 
 logger = getLogger(__name__)
@@ -113,5 +114,5 @@ class SpecialJudge:
             logger.info('$ %s', command)
             info, proc = utils.exec_command(command)
         if not self.is_silent:
-            logger.info(utils.NO_HEADER + 'judge\'s output:\n%s', utils.make_pretty_large_file_content(info['answer'] or b'', limit=40, head=20, tail=10, bold=True))
+            logger.info(utils.NO_HEADER + 'judge\'s output:\n%s', pretty_printers.make_pretty_large_file_content(info['answer'] or b'', limit=40, head=20, tail=10, bold=True))
         return proc.returncode == 0
