+++ onlinejudge_command/subcommand/generate_input.py
@@ -11,6 +11,7 @@ from logging import getLogger
 from typing import *
 
 import onlinejudge_command.format_utils as fmtutils
+import onlinejudge_command.pretty_printers as pretty_printers
 import onlinejudge_command.utils as utils
 
 logger = getLogger(__name__)
@@ -46,7 +47,7 @@ def write_result(input_data: bytes, output_data: Optional[bytes], *, input_path:
 
         if print_data:
             logger.info(utils.NO_HEADER + 'input:')
-            logger.info(utils.NO_HEADER + '%s', utils.make_pretty_large_file_content(input_data, limit=40, head=20, tail=10, bold=True))
+            logger.info(utils.NO_HEADER + '%s', pretty_printers.make_pretty_large_file_content(input_data, limit=40, head=20, tail=10, bold=True))
         with input_path.open('wb') as fh:
             fh.write(input_data)
         logger.info(utils.SUCCESS + 'saved to: %s', input_path)
@@ -54,7 +55,7 @@ def write_result(input_data: bytes, output_data: Optional[bytes], *, input_path:
         if output_data is not None:
             if print_data:
                 logger.info(utils.NO_HEADER + 'output:')
-                logger.info(utils.make_pretty_large_file_content(output_data, limit=40, head=20, tail=10, bold=True))
+                logger.info(pretty_printers.make_pretty_large_file_content(output_data, limit=40, head=20, tail=10, bold=True))
             with output_path.open('wb') as fh:
                 fh.write(output_data)
             logger.info(utils.SUCCESS + 'saved to: %s', output_path)
@@ -152,9 +153,9 @@ def try_hack_once(generator: str, command: str, hack: str, *, tle: Optional[floa
         expected = output_data.decode()
         if not simple_match(answer, expected):
             logger.info(utils.FAILURE + '' + utils.red('WA'))
-            logger.info(utils.NO_HEADER + 'input:\n%s', utils.make_pretty_large_file_content(input_data, limit=40, head=20, tail=10, bold=True))
-            logger.info(utils.NO_HEADER + 'output:\n%s', utils.make_pretty_large_file_content(answer.encode(), limit=40, head=20, tail=10, bold=True))
-            logger.info(utils.NO_HEADER + 'expected:\n%s', utils.make_pretty_large_file_content(output_data, limit=40, head=20, tail=10, bold=True))
+            logger.info(utils.NO_HEADER + 'input:\n%s', pretty_printers.make_pretty_large_file_content(input_data, limit=40, head=20, tail=10, bold=True))
+            logger.info(utils.NO_HEADER + 'output:\n%s', pretty_printers.make_pretty_large_file_content(answer.encode(), limit=40, head=20, tail=10, bold=True))
+            logger.info(utils.NO_HEADER + 'expected:\n%s', pretty_printers.make_pretty_large_file_content(output_data, limit=40, head=20, tail=10, bold=True))
             status = 'WA'
 
         if status == 'AC':
