+++ onlinejudge_command/subcommand/submit.py
@@ -8,6 +8,7 @@ from logging import getLogger
 from typing import *
 
 import onlinejudge_command.download_history
+import onlinejudge_command.pretty_printers as pretty_printers
 import onlinejudge_command.utils as utils
 
 import onlinejudge.dispatch as dispatch
@@ -52,7 +53,7 @@ def submit(args: 'argparse.Namespace') -> None:
 
     # report code
     logger.info('code (%d byte):', len(code))
-    logger.info(utils.NO_HEADER + '%s', utils.make_pretty_large_file_content(code, limit=30, head=10, tail=10, bold=True))
+    logger.info(utils.NO_HEADER + '%s', pretty_printers.make_pretty_large_file_content(code, limit=30, head=10, tail=10, bold=True))
 
     with utils.new_session_with_our_user_agent(path=args.cookie) as sess:
         # guess or select language ids
